Reduce exports, add comments, some renames
