+++ components/experimental/spaces/src/spacesComponentRegistry.ts
@@ -60,7 +60,7 @@ export class InternalRegistry implements IComponentRegistry, IComponentInternalR
     }
 }
 
-export const generateRegistryEntries = () => {
+const generateRegistryEntries = () => {
     // create a matching registry of type -> view type?  Import all view types above?
     // have to set the view registry on the dataModel though, so it can respond to getComponent() calls?
     const containerComponentsDefinition: IInternalRegistryEntry[] = [
